requirements.txt
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.2.0,<6.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
python-docx>=1.1.0,<2.0.0
reportlab>=4.1.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
google-genai>=1.0.0,<2.0.0
trafilatura>=2.0.0,<3.0.0
advertools>=0.16.0,<1.0.0
pytest>=8.2.0,<9.0.0
google-api-python-client>=2.170.0,<3.0.0
google-auth>=2.40.0,<3.0.0
google-analytics-data>=0.18.0,<1.0.0
plotly>=6.0.0,<7.0.0
kaleido>=0.2.1,<1.0.0
Django>=5.1.0,<6.0.0
ruff>=0.11.0,<1.0.0
pytest-django>=4.8.0,<5.0.0
coverage>=7.0.0,<8.0.0
